requirements.txt
<<<<<<< HEAD
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dj-rest-auth==7.0.1
Django==4.2.23
django-allauth==65.10.0
=======
asgiref==3.8.1
async-timeout==5.0.1
Django==4.2.13
>>>>>>> aa0d8d2aca20b0236fba9814baa464d1e69c63c0
django-dotenv==1.4.2
django-jazzmin @ git+https://github.com/farridav/django-jazzmin.git@d269874c96fe5ff6d231c3017cdd4650bc2f3105
django-querycount==0.8.3
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
<<<<<<< HEAD
flower==2.0.1
humanize==4.12.3
idna==3.10
inflection==0.5.1
kombu==5.5.4
packaging==25.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
=======
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.1.1
>>>>>>> aa0d8d2aca20b0236fba9814baa464d1e69c63c0
